+++ src/Controller/Admin/AppController.php
@@ -1,0 +1,69 @@
+<?php
+/**
+ * This file is part of me-cms.
+ *
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @copyright   Copyright (c) Mirko Pagliai
+ * @link        https://github.com/mirko-pagliai/me-cms
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
+ * @since       2.26.8
+ */
+namespace MeCms\Controller\Admin;
+
+use Cake\Event\Event;
+use MeCms\Controller\AppController as BaseAppController;
+
+/**
+ * Admin Application controller class
+ */
+class AppController extends BaseAppController
+{
+    /**
+     * Called before the controller action
+     * @param \Cake\Event\Event $event An Event instance
+     * @return \Cake\Network\Response|null
+     */
+    public function beforeFilter(Event $event)
+    {
+        parent::beforeFilter($event);
+
+        $this->viewBuilder()->setClassName('MeCms.View/Admin');
+
+        //Sets the paginate limit and the maximum paginate limit
+        //See http://book.cakephp.org/3.0/en/controllers/components/pagination.html#limit-the-maximum-number-of-rows-that-can-be-fetched
+        $this->paginate['limit'] = $this->paginate['maxLimit'] = getConfigOrFail('admin.records');
+
+        $this->Auth->deny();
+    }
+
+    /**
+     * Initialization hook method
+     * @return void
+     */
+    public function initialize()
+    {
+        parent::initialize();
+
+        $this->loadComponent('MeTools.Uploader');
+    }
+
+    /**
+     * Internal method to set an upload error.
+     *
+     * It saves the error as view var that `JsonView` should serialize and sets
+     *  the response status code to 500.
+     * @param string $error Error message
+     * @return void
+     * @since 2.18.1
+     */
+    protected function setUploadError($error)
+    {
+        $this->response = $this->response->withStatus(500);
+
+        $this->set(compact('error'));
+        $this->set('_serialize', ['error']);
+    }
+}