+++ src/Controller/Admin/SystemsController.php
@@ -17,7 +17,7 @@ use Cake\Core\Configure;
 use Cake\I18n\I18n;
 use Cake\Routing\Router;
 use League\CommonMark\CommonMarkConverter;
-use MeCms\Controller\AppController;
+use MeCms\Controller\Admin\AppController;
 use MeCms\Core\Plugin;
 use MeCms\Utility\Checkup;
 use Symfony\Component\Finder\Finder;