+++ src/Controller/Admin/UsersController.php
@@ -14,7 +14,7 @@ namespace MeCms\Controller\Admin;
 
 use Cake\Event\Event;
 use Cake\Mailer\MailerAwareTrait;
-use MeCms\Controller\AppController;
+use MeCms\Controller\Admin\AppController;
 use Symfony\Component\Finder\Finder;
 use Thumber\Utility\ThumbManager;